Add new text encoding phase.
This new phase takes the character values and maps them as-needed to
integers, so that we only have an output vocabulary size of as many
characters as we need.

Basically the way this works is, if you initialize it, and then pass
it:

"auto."

it'll give you:

[1, 2, 3, 4, 5]

Because it maps 'a' -> 1, 'u' -> 2, etc. Then, if you give it:

"ottowa"

it'll return:

[4, 3, 3, 4, 6, 1]

Because it already has a map for 'o' -> 4 and 't' -> 3, etc, but the
letter 'w' is new so it gets the next index available, 6.

Except instead of the indices starting at 1, they start at the value
of `num_reserved_tokens`, which is 2 as of writing. Because we want to
keep the tokens 0 and 1 reserved for indicating 'start of stream' and
'end of stream'.