Add run-test.sh script : use with caution!
run-test.sh is a script with the interface of queue-test.sh (allowing
multiple train+reports to be run at once, and creating unique temp
files for their inputs and outputs), but it doesn't use tsp, so tasks
are allowed to train in parallel. For the k-nearest-neighbor
predictor, this is fine, because the "training" phase (building the
search tree) is done on a single CPU thread. However, this script
should not be used for predictors which train on the graphics card, as
training multiple predictors on the graphics card will usually run out
of memory and trigger a cuda assert