Tokenizer freezing
This system allows you to "freeze" the tokenizer once all training
samples have entered it, so that any new unrecognized characters are
assigned an "Unknown" token, instead of being given a fresh
token. This ensures that the vocabulary size doesn't change after
training, so that all the token vectors can line up.