Add more flexible data filters
System in place to control predicates on which data pairs are trained
on via the command line. Default is currently all except ones where
the expected tactic is a compound command or a bullet. Also added
"goal-changes" filter, which only allows contexts where the "right"
tactic changed the goal in some way.