Implement percentage-pass.py script
This script lets you specify a filter, and some csv files, and will
tell you how many tactics in the csv files pass the filter, and how
many proofs are entirely made up of only tactics which pass the
filter. WARNING: tactics filtered in the original report generating
the csvs don't show up as tactics in this phase.