Upgrade context filter DSL
The context filter DSL now supports the infix binary operators '%' for
AND and '+' for OR, but explicit association with parenthesis is
required when mixing operators. There are also the new meta-filters
"tactic:*" and "etactic:*", which selects contexts with a certain
tactic as their prefix. For instance "tactic:intros" will select
contexts with the tactic "intros." or "intros a b." The "etactic:*"
variation also accepts the tactic prefixed with an "e", so
"etactic:auto" accepts both "eauto." and "auto".