Smarter "relevance" tokenizer
The old relevance tokenizer worked as such:

For each potential keyword, measure the information gain of splitting
samples on it. That is, split the samples into two pools, those with
the word, and those without, and average their entropies. Then, pick
the top k keywords based on that measure.

The problem with this system is that it tends to pull out highly
correlated keywords, because each keyword is judged
independently.

The new system does a similar thing, but picking the keywords one at a
time, and conditioning on past keywords for future ones. So, the first
keyword is picked based on the information gain metric above. But
then, the samples are split into pools based on whether they have the
first keyword. The entropy of the pools is measured, and the one with
the highest entropy is used to pick the next keyword, which splits all
the pools again, etc.

This appears to perform much better.