Handle persistent anomalies
Previous approach to handling anomalies was to restart the coq
process, fast-forward through previous proofs using found solutions,
and try again. However, when the anomaly is caused by, say, a
`compute` invocation during search always timing out and not allowing
interrupt, this is problematic, as we'll just have to hit the anomaly
twice and give up, when we could just give up on the proof. Now, we'll
just give up on the proof and move on (with a restarted instance).