Make output of file scrapes reflect partiality
To aid in this script being used for Makefiles where the .scrape files
are the primary target, scrapes that are interrupted will instead
write .scrape.partial files, and .scrape files will indicate that the
scrape was completed. Thus, .scrape files can reliably be used as a
dependency in makefile targets.