Add support for weights "preludes"
Some sets of weightsfile, because they were trained on a particular
project, are going to have specific tactics they invoke often, which
may not be available in other projects.

For instance, CompCert has an "inv" tactic, defined as "inv h :=
inversion h; clear h; subst.". It is used often, in place of
inversion. When running using the compcert weights on other projects,
however, the "inv" tactic isn't necessarily defined.

To better facilitate cross-project usage, this commit introduces a
"prelude" file attached to trained weights. If the weights file is
called "foo.dat", then "foo.prelude.v", if it exists, is loaded and
every command in it is run before searching a file using "foo.dat".

In the Compcert example, this allows us to put the definition of "inv"
in the prelude, and use that tactic on other projects.