Add --read-progress/-p flag for showing the progress of reading file
When I'm looking for detailed information about a failure, I usually
use -vv in favor of -P to see the commands that are actually getting
run. However, for some larger files, this can result in a whole lof of
nothing for a while, while the file commands are being parsed. The new
-p flag will show a progress bar for reading the file, and then will
allow -vv to do it's thing.