Make sampler return batch size items more often, and refactor train
This pulls the train method apart into several functions. One which samples 'fairly' from the two buffers, one for computing the outputs of each type of sample, and finally one that calls those and trains.

The fair sampling behaves the same as before if both buffers have a lot of items; it splits the batch size into two equal parts, and samples from each buffer with those sizes. But, if one of the buffers doesn't have enough items to fulfill it's half, now instead of just returning fewer samples, we try to make it up in the other buffer. This results in full batches more of the time.